chore: change module structure for connectors
