+++ examples/s2s/conversational.py
@@ -25,8 +25,7 @@ from langchain_core.callbacks import BaseCallbackHandler
 from langchain_core.messages import AIMessage, HumanMessage, SystemMessage
 from rai.agents.base import BaseAgent
 from rai.communication import BaseConnector
-from rai.communication.ros2.api import IROS2Message
-from rai.communication.ros2.connectors import ROS2HRIConnector, TopicConfig
+from rai.communication.ros2 import IROS2Message, ROS2HRIConnector, TopicConfig
 from rai.utils.model_initialization import get_llm_model
 
 from rai_interfaces.msg import HRIMessage as InterfacesHRIMessage
