+++ src/rai_core/rai/communication/ros2/connectors/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Robotec.AI
+# Copyright (C) 2025 Robotec.AI
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,26 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .api import (
-    ConfigurableROS2TopicAPI,
-    IROS2Message,
-    ROS2ActionAPI,
-    ROS2ServiceAPI,
-    ROS2TopicAPI,
-    TopicConfig,
-)
-from .connectors import ROS2ARIConnector, ROS2HRIConnector
-from .messages import ROS2ARIMessage, ROS2HRIMessage
+from .ari_connector import ROS2ARIConnector
+from .hri_connector import ROS2HRIConnector
 
 __all__ = [
-    "ConfigurableROS2TopicAPI",
-    "IROS2Message",
     "ROS2ARIConnector",
-    "ROS2ARIMessage",
-    "ROS2ActionAPI",
     "ROS2HRIConnector",
-    "ROS2HRIMessage",
-    "ROS2ServiceAPI",
-    "ROS2TopicAPI",
-    "TopicConfig",
 ]
