+++ src/rai_core/rai/tools/ros2/services.py
@@ -24,7 +24,7 @@ from typing import Any, Dict, Type
 from langchain_core.tools import BaseTool
 from pydantic import BaseModel, Field
 
-from rai.communication.ros2.connectors import ROS2ARIConnector, ROS2ARIMessage
+from rai.communication.ros2 import ROS2ARIConnector, ROS2ARIMessage
 
 
 class CallROS2ServiceToolInput(BaseModel):
