+++ src/rai_sim/rai_sim/o3de/o3de_bridge.py
@@ -23,7 +23,7 @@ from typing import Any, Dict, List, Optional, Set
 import yaml
 from geometry_msgs.msg import Point, PoseStamped, Quaternion
 from geometry_msgs.msg import Pose as ROS2Pose
-from rai.communication.ros2.connectors import ROS2ARIConnector, ROS2ARIMessage
+from rai.communication.ros2 import ROS2ARIConnector, ROS2ARIMessage
 from rai.utils.ros_async import get_future_result
 from std_msgs.msg import Header
 from tf2_geometry_msgs import do_transform_pose
