+++ tests/communication/ros2/test_connectors.py
@@ -19,7 +19,7 @@ import pytest
 from PIL import Image
 from pydub import AudioSegment
 from rai.communication import HRIPayload
-from rai.communication.ros2.connectors import (
+from rai.communication.ros2 import (
     ROS2ARIConnector,
     ROS2ARIMessage,
     ROS2HRIConnector,
